lambda_functions/stream_records/requirements.txt
requests
protobuf
aws-kinesis-agg
amazon-ion==0.7.0